Move simple control implementations to dedicated dir
