+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from './aspect';
 export * from './container';
 export * from './control';
+export * from './controls';
 export * from './conversion';
 export * from './converter';
 export * from './data';
@@ -12,5 +13,4 @@ export * from './namespace-aliaser.aspect';
 export * from './render-scheduler.aspect';
 export * from './style';
 export * from './submit.aspect';
-export * from './value';
 export * from './validation';
