+++ src/style/css-classes.aspect.ts
@@ -14,7 +14,8 @@ import {
   EventKeeper,
   EventNotifier,
   eventSupply,
-  EventSupply, eventSupplyOf,
+  EventSupply,
+  eventSupplyOf,
   isEventKeeper,
   nextAfterEvent,
   trackValue,
